so-middleware/load-app/requirements.txt
kafka-python==2.0.2
pandas==2.1.4
python-dotenv==1.0.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-instrumentation-kafka-python==0.42b0
opentelemetry-instrumentation-requests==0.42b0
requests==2.31.0